Fix jQuery UI's dialog placement in RTL
jQuery UI uses 'left' to position dialogs which perhaps isn't the
the perfect approach for RTL but more or less has worked.

The styles shipped with jQuery UI also feature a 'left: 0' which
immediately is replaced by the JavaScript side for each dialog
but the problem start when CSS Janus turns that to 'right: 0'
and that's why local wikis had to apply 'right: auto' in their local
styles (see the survey in T344467) just to fix dialog compatibility
with RTL and this fix adds a @noflip to let CSS Janus know it
shouldn't turn this one to 'right: 0' and the fix should be
compatible with workarounds RTL wikis have applied locally.

Bug: T86979
Bug: T344467
Bug: T361463
Change-Id: Ia48990a7f367b772905d70ccab3b2ca5a9f4e220