Use base font size for wikitext editors
We don't want to filter out action=edit/submit as the
page also can contain a content preview, which should
respect the appearance preference.

Instead reset the UI size for mw-editform.

Note that the font size of the edit are is already fixed
at 13px, so this just fixes the size of UI components,
which shouldn't be scaled by the appearance setting.

Bug: T373875
Change-Id: Iaabdb67da23b5722056046c1ed88e366471d7b2c