Remove PerformanceBudgetTest
This test has not served the original purpose it intended and is
now more of a nuisance than supportive to our goal of monitoring
performance across MediaWiki.

For now, we are going to remove this test in favor of a new test
that we are currently finalizing in T360590 that we hope to land
after the next MediaWiki release.

Bug: T373017
Change-Id: I0eb22ebbed67ded008de7ca75f9f8f399a214ef8