styles: Remove barely visible special opacities in Vector legacy
The differences are barely visible for the human eye. And
`@opacity-icon-base` is only used in MenuDropdown once, for the rare
case of icons accompanying a menu item.
We shied away from using the Codex defaults back then as the values were
slightly different and the team was heavily occupied from getting
Vector 2022 off the ground.

Change-Id: Ia36402165b1e1185221f0b7bf865488cf7588ad5