styles: Harden design token usage in calculations by using `calc()`
As design token values could be numeric values with units,
but also –with introduction of font modes in Codex[1]–
CSS custom properties, it's useful to rely on CSS `calc()`
function to ensure that the values are correctly calculated.
This patch will still be properly computed these values once
the font modes work is merged.
This will also bring us closer to removing Less parser
in future and rely on CSS standard only.

[1] https://gerrit.wikimedia.org/r/c/design/codex/+/1123503

Bug: T389116
Bug: T389234
Co-Authored-by: Anne Tomasevich <atomasevich@wikimedia.org>
Change-Id: I11181d9f416c6a001d5bbf5baf099315fff1a528