clientPreferences: Make appearance buttons follow current Codex markup
This applies to Codex Radios and ToggleSwitches. The markup got out of
sync over the many months since the appearance menu got started. With
the upcoming changes for font modes, this leads to `line-height`
discrepancies in the radio buttos.

This patch brings the markup back in sync by adding a child `<span>` to the `.cdx-label` element.

Bug: T390246
Change-Id: I447c2b0af6e296b06687b00dcef28ed40d5d4be9