Separate desired heading scroll offset from sticky header height
Per T314419, when a section link is clicked in the table of contents,
Vector brings the heading into a position that is easier to read,
which is slightly lower than the very top of the page.

This was implemented using scroll-padding-top on the document root,
which unfortunately also applied to other elements where this behavior
is not desirable, for example Cite backlinks (T384777). Elements that
are not as prominent as headings should be aligned with the very top,
to make them easier to spot.

Instead, implement it using scroll-margin-top on the headings,
and use scroll-padding-top only for the sticky header offset.

(Remove unused variable in skins.vector.styles.legacy.)

Bug: T364023
Bug: T384777
Change-Id: I759e34aabbb44e7fb0e4e8593529581076aba67b