styles: Workaround for Codex 2.0 release
The upcoming Codex 2.0 release changes the name of a mixin
file used in Vector. This patch is needed to get around a
"chicken-and-egg" situation where the patch in MW core
(which introduces the new Codex files, including one with
a changed name) fails tests because Vector expects the old
file to still be present.

This patch temporarily comments-out the offending code in Vector
in a way that should not have any user-facing impact (because
all the light-mode values are provided as fallbacks).

A follow-up patch in Vector after the Codex 2.0 update in Core
will undo these changes and update the Less import to the new
filename.

Bug: T391012
Change-Id: Ia5cbdb6b1c9e538248a6ea23705cbc49fc585471