styles: Exclude Vector page titlebar & toolbar from user font size pref
This commit is in preparation for Codex font modes rollout. It
excludes the entire Vector's page titlebar and page toolbar from the
user's Vector font size preferences.  It doesn't change anything in
status quo of the behavior, but it will allow us to introduce Codex
font modes without visual impact.

Bug: T393033
Change-Id: Ia56191a3c69d9ee4744aa347927619a0e7ffb41d