styles: Set override also to former value of `line-height-small` token
This change is a stop-gap measure to ensure that the line height
specific skin overrides are applied with the former values in
preparation of new font modes.
Related follow-up to I0ec219861711.

These variable overrides can be removed once the new font modes
typographic styles and mixins are fully implemented.

Bug: T389900
Bug: T394305
Change-Id: Ic302a076f4ec0224a44faa4746112e3cb5a77153